% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{Capitalize first letter}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{character string to be capitalized}

\item{strict}{TRUE forces everything to 1st letter capital; FALSE allows for things like acronyms to remain all caps; Defaults to TRUE.}
}
\value{
s_cap Original string with first letter capitalization
}
\description{
Capitalizes first letter of a string. Copied from https://stat.ethz.ch/R-manual/R-devel/library/base/html/chartr.html
}
\examples{
plot_crayons()

}

