% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{cb_pal}
\alias{cb_pal}
\title{Color Blind Palette}
\usage{
cb_pal(n, pal = "wong", sort = TRUE)
}
\arguments{
\item{n}{numeric number of colors to generate}

\item{pal}{character string of color palette}

\item{sort}{TRUE sorts colors based on their distances in HCL space}
}
\value{
pal_result vector color palette in HEX
}
\description{
Pulls together some of the best or more common palettes for R, including those focused on colorblindness.
wong palette is based on Wong, Bang. "Points of view: Color blindness." nature methods 8.6 (2011): 441-441.
few palette is based on qualitative color palettes from Stephen Few, http://www.perceptualedge.com/articles/visual_business_intelligence/rules_for_using_color.pdf.
ptol palette is based on qualitative color palettes from Paul Tol, https://personal.sron.nl/~pault/colourschemes.pdf.
d3 palette is based on d3 standard. nature is based on Nature Publishing Group
}

