% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legend.R
\name{extract_legend}
\alias{extract_legend}
\title{Extract Legend}
\usage{
extract_legend(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplot2 figure}
}
\value{
legend legend extracted from ggplot2
}
\description{
Extracts the legend from a ggplot object
}
