% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_percent_label.R
\name{ggplot_percent_label}
\alias{ggplot_percent_label}
\title{ggplot Percent Labels}
\usage{
ggplot_percent_label(theargument, siglevel = 3)
}
\arguments{
\item{theargument}{label for ggplot}

\item{siglevel}{significant figures}
}
\value{
label for ggplot2
}
\description{
This function allows for custom percentage labels in ggplot2.
}
