% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_bootstrap.R
\name{gof_bootstrap}
\alias{gof_bootstrap}
\title{Goodness of Fit for Univariate Distributions with Bootstrapping}
\usage{
gof_bootstrap(fit, n_sims = 50000, parallel = FALSE)
}
\arguments{
\item{fit}{fitdistrplus object}

\item{n_sims}{number of replications, defaults to 5e4 (50,000)}

\item{parallel}{TRUE or FALSE option to run in parallel}
}
\value{
gof_result list with KS, AD, and CVM test statistic and p-value
}
\description{
Calculates a number of goodness of fit statatistics for a chosen univariate distribution. This function was created because published p-values for the Kolmogorov-Smirnov, Anderson-Darling and Cramer-von-Mises tests are not valid if the distribution is estimated from the data. This function uses bootstrapping to obtain the p-value based on	the distribution and parameters.
}
