% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_bootstrap.R
\name{gof_replicate}
\alias{gof_replicate}
\title{Goodness of Fit replicates}
\usage{
gof_replicate(i, param_replicate, n, distr_name)
}
\arguments{
\item{i}{replication index}

\item{param_replicate}{parameters for probability distribution function}

\item{n}{number of observations to use in each replication}

\item{dist_name}{name of univariate distribution}
}
\value{
gof_results list with KS, AD, and CVM test statistic
}
\description{
Wrapper to generate random samples from a probability distribution and run gof_stats
}
