% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_ts.R
\name{gof_ts}
\alias{gof_ts}
\title{Goodness of Fit for a time series}
\usage{
gof_ts(pred, obs)
}
\arguments{
\item{pred}{vector of predicted values}

\item{obs}{vector of observed values}
}
\value{
gof_results list of goodness of fit statistics
}
\description{
Calculates a number of goodness of fit statatistics for a time series. These include, Mean Error (ME), Mean Absolute Error (MAE), Mean Squared Error (MSE), Root Mean Squared Error (RMSE), Nash-Sutcliffe Efficiency (NSE), Pearson Correlation, and Spearman Correlation.
}
