% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_func.R
\name{pca_calc}
\alias{pca_calc}
\title{PCA calculation with centering and scaling}
\usage{
pca_calc(data)
}
\arguments{
\item{data}{a matrix with input data}
}
\value{
pca_list Output list, with loadings, center, scaling, and importance.
}
\description{
Performs Principal Component Analysis (PCA) with centering and scaling
}
\examples{
pca_calc()

}
