% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_cum_var_plot}
\alias{pca_cum_var_plot}
\title{PCA Cumulative Variance Explained Plot}
\usage{
pca_cum_var_plot(importance)
}
\arguments{
\item{importance}{a dataframe with PCA importance data}
}
\value{
p a plot of Cumulative Variance explained
}
\description{
Plots Cumulative Variance explained from PCA analysis
}
