% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_eigen_plot}
\alias{pca_eigen_plot}
\title{PCA Eigen Plot}
\usage{
pca_eigen_plot(importance)
}
\arguments{
\item{importance}{a dataframe with PCA importance data}
}
\value{
p an eigen value plot
}
\description{
Plots Eigen values from PCA analysis
}
