% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_func.R
\name{pca_reconstruct}
\alias{pca_reconstruct}
\title{PCA reconstruction using loading}
\usage{
pca_reconstruct(data, pca_fit)
}
\arguments{
\item{data}{a matrix with input data}

\item{pca_fit}{a PCA object from the function pca_calc}
}
\value{
x A time series of PCA scores.
}
\description{
Calculates Principal Component Analysis (PCA) scores from a loading matrix and new data.
}
\examples{
pca_reconstruct()

}
