% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_var_plot}
\alias{pca_var_plot}
\title{PCA Proportion of Variance Explained Plot}
\usage{
pca_var_plot(importance)
}
\arguments{
\item{importance}{a dataframe with PCA importance data}
}
\value{
p a plot of Proportion Variance explained
}
\description{
Plots Proportion of Variance explained from PCA analysis
}
