% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{separate_qual_colors}
\alias{separate_qual_colors}
\title{Order Qualitative Colors}
\usage{
separate_qual_colors(pal, preserve_first = TRUE)
}
\arguments{
\item{pal}{vector of HEX colors}

\item{preserve_first}{TRUE preserve the first color in the vector.}
}
\value{
color_separated vector of colors separated by HCL
}
\description{
Orders qualitative colors by finding the largest difference between each color based on HCL
}
