% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift a time series}
\usage{
shift(x, shift_by)
}
\arguments{
\item{x}{vector timeseries}

\item{shift_by}{numeric number of spaces to shift. Can be positive or negative.}
}
\value{
out resulting numeric vector
}
\description{
Copied from https://www.r-bloggers.com/generating-a-laglead-variables/
}
