% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_bylag.R
\name{shift_bylag}
\alias{shift_bylag}
\title{Shift a time series by a given lag}
\usage{
shift_bylag(x, shift_by)
}
\arguments{
\item{x}{vector timeseries}

\item{shift_by}{numeric number of spaces to shift. Can be positive or negative.}
}
\value{
out resulting numeric vector
}
\description{
Copied from https://www.r-bloggers.com/generating-a-laglead-variables/
}
