% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic_new.R
\name{theme_classic_new}
\alias{theme_classic_new}
\title{Theme Classic New}
\usage{
theme_classic_new(legend_pt = 10)
}
\arguments{
\item{legend_pt}{font size}
}
\value{
theme_out theme for ggplot2
}
\description{
An updated version of theme_classic from ggplot2 using my personal preferences. Had to create this due to a previous error in ggplot2.
}
