% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic_new.R
\name{theme_classic_new_majgrid}
\alias{theme_classic_new_majgrid}
\title{Theme Classic New with Major Gridlines}
\usage{
theme_classic_new_majgrid(legend_pt = 10)
}
\arguments{
\item{legend_pt}{font size}
}
\value{
theme_out theme for ggplot2
}
\description{
An updated version of theme_classic from ggplot2 using my personal preferences with major gridlines
}
