% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_daily_dl.R
\name{usgs_daily_dl}
\alias{usgs_daily_dl}
\title{Download daily data from USGS}
\usage{
usgs_daily_dl(site_id, parameterCd = "00060", dest_folder = getwd())
}
\arguments{
\item{site_id}{USGS site id}

\item{dest_folder}{location to save file}

\item{param_cv}{USGS parameter id}
}
\value{
dl_result a character stating "Success" or "Failure"
}
\description{
Use the dataRetrieval package to download daily stream flows.
}

