% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_month_dl.R
\name{usgs_month_dl}
\alias{usgs_month_dl}
\title{Download monthly data from USGS}
\usage{
usgs_month_dl(site_id, parameterCd = "00060", destination_folder = getwd())
}
\arguments{
\item{site_id}{USGS site id}

\item{destination_folder}{location to save file}

\item{param_cv}{USGS parameter id}
}
\value{
dl_result a character stating "Success" or "Failure"
}
\description{
The current dataRetrieval package does not allow for download of monthly stats.  This function downloads this file for read-in later.
}
