% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_readin.R
\name{usgs_readin}
\alias{usgs_readin}
\title{Read in saved data from USGS}
\usage{
usgs_readin(site_id, param_cd = "00060", time_param = "daily",
  destination_folder = getwd())
}
\arguments{
\item{site_id}{USGS site id}

\item{time_param}{time accumulation}

\item{destination_folder}{location to save file}

\item{param_cv}{USGS parameter id}
}
\value{
dl_result a character stating "Success" or "Failure"
}
\description{
Reads in local USGS data file using dataRetreival package.
}
