% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_wateryear.R
\name{usgs_wateryear}
\alias{usgs_wateryear}
\title{Determine water year}
\usage{
usgs_wateryear(year, month, first_month = 10)
}
\arguments{
\item{year}{USGS year of observation}

\item{month}{USGS month of observation}

\item{first_month}{first month of water year, assumes USGS default as Oct 1}
}
\value{
water_year water year of observation
}
\description{
Reads in a year and month and calculates water year.
}
