% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_wateryear.R
\name{usgs_wateryear_inverse}
\alias{usgs_wateryear_inverse}
\title{Inverse of water year calculation}
\usage{
usgs_wateryear_inverse(water_year, month, first_month = 10)
}
\arguments{
\item{water_year}{water year of observation}

\item{month}{USGS month of observation}

\item{first_month}{first month of water year, assumes USGS default as Oct 1}
}
\value{
year year of observation
}
\description{
Reads in a water year and month and calculates the year.
}
