% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap Text}
\usage{
wrap_text(text, length = 30)
}
\arguments{
\item{text}{original text}

\item{length}{maximum number of characters per row}
}
\value{
final_text text with line breaks every length-th character
}
\description{
Reads in text and splits into separate lines
}
